--- trustrag/modules/document/resume/entities/res/corp.tks.freq.json
@@ -1,65 +1,0 @@
-[
-        "科技",
-        "集团",
-        "网络科技",
-        "技术",
-        "信息",
-        "分公司",
-        "信息技术",
-        "发展",
-        "科技股份",
-        "网络",
-        "贸易",
-        "商贸",
-        "工程",
-        "企业",
-        "集团股份",
-        "商务",
-        "工业",
-        "控股集团",
-        "国际贸易",
-        "软件技术",
-        "数码科技",
-        "软件开发",
-        "有限",
-        "经营",
-        "科技开发",
-        "股份公司",
-        "电子技术",
-        "实业集团",
-        "责任",
-        "无限",
-        "工程技术",
-        "上市公司",
-        "技术开发",
-        "软件系统",
-        "总公司",
-        "网络服务",
-        "ltd.",
-        "technology",
-        "company",
-        "服务公司",
-        "计算机技术",
-        "计算机软件",
-        "电子信息",
-        "corporation",
-        "计算机服务",
-        "计算机系统",
-	"有限公司",
-	"事业部",
-	"公司",
-	"股份",
-	"有限责任",
-	"软件",
-	"控股",
-	"高科技",
-	"房地产",
-	"事业群",
-	"部门",
-	"电子商务",
-	"人力资源顾问",
-	"人力资源",
-	"株式会社",
-	"网络营销"
-]
-
